Version 1.2.0
1. Deprecated obj.type

`maybe()` function now checks if an object is a `maybe` istance with
`instanceof`. `obj.type === 'maybe'` wasn't a great idea, and
this property will be removed in a future release.

2. Deprecated `maybe()` second argument

The second argument was designed to change the default
emptiness definition. That wasn't a great idea either
and actually I never used this feature. It will be removed too.

3. `maybe.object()`

Now checks if the object is a `maybe` instance like `maybe()`.